[core] add logic to unregister modules on teardown
[core|linux] implement unregister
[core] call unregister in separate pass on teardown

we first unregister, then destroy modules. we do this because we want to
make sure that modules can still depend on other modules for while
unregistering (le_log, for example)

[core] only load module if not found in cache

firt test in cache if a requested module already has been loaded, if so,
just return cached api

[core] refactor module load macros

add a means to explicitly load a module api

[core|linux] call logger api explicitly

don't rely on references for calling the logger api - as we don't trust
the compiler to update these correctly

[log] make logging more robust

enforce that api for logging is available - we do this by first testing
whether the compilation-unit-local static cache of the api has been
initialized. if not, we request the api explicitly.

this is so that we can log from nearly anywhere, especially in static
builds where static initialization order would otherwise bite us...

[debug_print_text] add cleanup on teardown

destroy singleton object when tearing down debug text printer

[le_png] update le_png

pull in latest version of le_png which removes global static le_logger
objects in favour of local statics, which are lazily initialized.

[*] enforce lazy initialization for logger objects

[le_debug_print_text] print log messages in debug

[le_pipeline] apply lazy le_log

[le_swapchain] use lazy logger

[le_window] lazy logger initializatino

[pipeline_builder,swapchain] lazy logger initialization

[ci] cleanup ci script