[img_swapchain] fix buffer allocation issue
images allocated with optimal tiling may occupy less bytes (or perhaps
even more) than a raw image of the same extents and pixel type.

we can therefore not assume that the buffer used to export the image has
the same size as the image's allocation;

we must calculate the buffer size manually, given image format and
extents.